refactor: reorganize fetch logic, improve with generics, and remove console logs
